--- src/components/api/SubstitutionReferenceContext.js
@@ -1,7 +1,0 @@
-/**
- * For communication between the "Try it" buttons and the REPL
- * on the Substitution Reference page
- */
-import React from 'react'
-
-export default React.createContext({})
